+++ src/market_sentinel_ai/ports/__init__.py
@@ -3,6 +3,7 @@ from market_sentinel_ai.ports.backtesting import BacktestEngine
 from market_sentinel_ai.ports.features import FeatureEngine
 from market_sentinel_ai.ports.market_data import MarketDataProvider
 from market_sentinel_ai.ports.models import PredictiveModel
+from market_sentinel_ai.ports.order_book import OrderBookProvider
 from market_sentinel_ai.ports.reasoning import ReasoningProvider
 from market_sentinel_ai.ports.storage import CandleRepository
 
@@ -12,7 +13,7 @@ __all__ = [
     "CandleRepository",
     "FeatureEngine",
     "MarketDataProvider",
+    "OrderBookProvider",
     "PredictiveModel",
     "ReasoningProvider",
 ]
-